[NA] minor fixes for guardrails (#1925)
* change GuardrailBatchItem to JSON views

* rename GuardrailBatchItem to Guardrail

* rename addGuardrailsBatch to createGuardrails

* removed unnecessary validation

* changed doubles to floats

* spotless

* fix compilation error

+++ apps/opik-backend/src/main/java/com/comet/opik/api/GuardrailPiiDetails.java
@@ -16,6 +16,6 @@ public record GuardrailPiiDetails(Map<String, List<Item>> detectedEntities) {
     @Builder(toBuilder = true)
     @JsonIgnoreProperties(ignoreUnknown = true)
     @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-    public record Item(double score) {
+    public record Item(float score) {
     }
 }
